Fixing for resubmission after review
Here's the feedback from Doris Keller when I pushed this in the review-my-project channel.

:meets: Only one <form> tag is being used to contain all the form elements.
:meets: Fieldsets and legends have been added for each of the following sections: ‘Contact Information’ and ‘Newsletter’ sections
:meets: All required form fields are present. Text, email, telephone, select, checkboxes, radio buttons, textarea, and submit button.
:meets: Each field includes the necessary attributes. ID, type, and name for inputs. Select/textarea elements have id and name attributes.
:meets: Only one radio button can be selected at a time.
:meets: Full name and email address fields have “required” placeholder text.
:meets: The HTML file include the viewport meta tag in the head of each document.
:meets: A mobile-first approach is utilized using min-width properties for media queries.
:meets: Appropriate media query is in place to match the screen size for tablet and desktop screens at 768px.
:meets: The label text position matches the mockups: above the form field for mobile and to the left of the form field for desktop
:meets: Project uses Flexbox for layout.
:meets: Uses a Google Font.
:meets: Includes focus states for text fields, select menu, and the textarea field.
:meets: HTML and CSS code passes the validation.
:meets: HTML and CSS errors are accepted exceptions.
:meets: :exceeds: Placeholder text has been added to all text inputs and the textarea element.
:meets: :exceeds: Uses the required attribute to add HTML5 Validation for required fields.
:needs-work: Each form field includes a label, including the textarea element, and each label’s `for` attribute is associated with a unique `id` on its corresponding form field.
>
All of your form fields are having a label - with one exception: Make sure to use a <label> instead of the paragraph for your textarea too.
:needs-work: Zip code width should be 50% width of the other input fields in mobile view and 25% in tablet view.
>
Unluckily, you’re not quite there yet. It looks almost perfect in mobile view, but there are still some issues in desktop view - see screenshot below.
:needs-work: Roughly matches layout of the mobile and desktop mockups. General spacing and arrangement of the elements matches the design of the mockups for mobile and desktop
>
Almost perfect! But currently the select field is sitting besides the zip code field instead of above it. It also should have the same width as the other input fields.
The problem lays on how you set up the div wrapping the select field and the zip code fields as well as the shared value of 50% of both divs inside this div - line 322 of your css.
:needs-work: :exceeds: BOTH of the following additional styling enhancements have been added to the form: change the background color for one of the main sections and added CSS transitions for focus states.
>
Well done for the focus-states on input fields!
I can find transitions on labels, but not on the fields. Try to add those too. I’m not sure if targeting the labels on focus has any impact at all.
After fixing those small issues your amazing looking form should be ready for an exceeds! :fire: